+++ src/components/Admin/ParentAdmin.jsx
@@ -35,7 +35,6 @@ import {
 } from '../../utils/imageFallback';
 import GotifySettings from './GotifySettings';
 import HomeAssistantSettings from './HomeAssistantSettings';
-import NextcloudSettings from './NextcloudSettings';
 import ProblemReportsPanel from './ProblemReportsPanel';
 import WebPushSettings from './WebPushSettings';
 
@@ -515,7 +514,6 @@ export default function ParentAdmin({ onOpenFamilyTree }) {
         <WebPushSettings />
         <GotifySettings />
         <HomeAssistantSettings />
-        <NextcloudSettings />
         <ProblemReportsPanel />
 
         <section className="admin-panel admin-media-panel">